Use custom data structures to simplify card draw logic further
